+++ packages/frontend/src/components/common/Table/ScrollableTable/index.tsx
@@ -1,4 +1,6 @@
+import { FeatureFlags } from '@lightdash/common';
 import { useRef, type FC } from 'react';
+import { useFeatureFlagEnabled } from '../../../../hooks/useFeatureFlagEnabled';
 import { Table, TableScrollableWrapper } from '../Table.styles';
 import { useTableContext } from '../useTableContext';
 import TableBody from './TableBody';
@@ -8,17 +10,25 @@ import TableHeader from './TableHeader';
 interface ScrollableTableProps {
     minimal?: boolean;
     showSubtotals?: boolean;
+    isDashboard?: boolean;
 }
 
 const ScrollableTable: FC<ScrollableTableProps> = ({
     minimal = true,
     showSubtotals = true,
+    isDashboard = false,
 }) => {
     const { footer } = useTableContext();
     const tableContainerRef = useRef<HTMLDivElement>(null);
+    const isDashboardRedesignEnabled = useFeatureFlagEnabled(
+        FeatureFlags.DashboardRedesign,
+    );
 
     return (
-        <TableScrollableWrapper ref={tableContainerRef}>
+        <TableScrollableWrapper
+            ref={tableContainerRef}
+            $isDashboard={isDashboard && isDashboardRedesignEnabled}
+        >
             <Table $showFooter={!!footer?.show}>
                 <TableHeader minimal={minimal} showSubtotals={showSubtotals} />
                 <TableBody