+++ packages/frontend/src/features/dashboardHeader/FilterGroupSeparator.tsx
@@ -5,20 +5,29 @@ import MantineIcon from '../../components/common/MantineIcon';
 
 type Props = {
     icon: Icon;
-    tooltipLabel: JSX.Element;
+    tooltipLabel?: JSX.Element;
+    size?: 'md' | 'sm';
 };
 
-const FilterGroupSeparator: FC<Props> = ({ icon, tooltipLabel }) => {
+const FilterGroupSeparator: FC<Props> = ({
+    icon,
+    tooltipLabel,
+    size = 'md',
+}) => {
     return (
-        <Tooltip label={tooltipLabel}>
+        <Tooltip disabled={!tooltipLabel} label={tooltipLabel}>
             <Center
-                w={30}
-                h={30}
+                w={size === 'md' ? 30 : 24}
+                h={size === 'md' ? 30 : 24}
                 flex="0 0 auto"
                 bg="ldGray.1"
                 style={{ borderRadius: '100%' }}
             >
-                <MantineIcon icon={icon} color="ldGray.6" />
+                <MantineIcon
+                    icon={icon}
+                    color="ldGray.6"
+                    size={size === 'md' ? undefined : 16}
+                />
             </Center>
         </Tooltip>
     );